# src/presentation/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1",
    "title": "",
    "description": ""
  },
  "paths": {
...
    "/user/publicData": {
      "get": {
        "operationId": "getManyUsersPublicData",
        "description": "",
        "tags": ["userData"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "searchNick",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPublicData"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/privateData": {
      "get": {
        "operationId": "getManyUsersPrivateData",
        "description": "ADMIN role requred. Get private users data.",
        "tags": ["userData", "ADMIN ONLY"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "searchNick",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}": {
      "put": {
        "operationId": "updateUser",
        "description": "",
        "tags": ["userData"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/completeExercise/{exerciseId}": {
      "put": {
        "operationId": "updateUser",
        "description": "",
        "tags": ["userData"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/publicData": {
      "get": {
        "operationId": "getOneUserPublicData",
        "description": "",
        "tags": ["userData"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicData"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}/privateData": {
      "get": {
        "operationId": "getOneUserPrivateData",
        "description": "ADMIN role requred. Get one user data.",
        "tags": ["userData", "ADMIN ONLY"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ROLE": {
        "type": "string",
        "enum": ["ADMIN", "USER"]
      },
...
      "User": {
        "properties": {
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "age": {
            "type": "number"
          },
          "role": {
            "$ref": "#/components/schemas/ROLE"
          },
          "completedExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedExercise"
            }
          }
        },
        "required": ["name", "surname", "nickName", "email", "age", "role"]
      },
      "UserPublicData": {
        "properties": {
          "id": {
            "type": "number"
          },
          "nickName": {
            "type": "string"
          }
        },
        "required": ["name", "nickName"]
      },
      "CompletedExercise": {
        "properties": {
          "exerciseId": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          }
        }
      },
...
      "updateUser": {
        "properties": {
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "nickName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "age": {
            "type": "number"
          },
          "role": {
            "$ref": "#/components/schemas/ROLE"
          }
        }
      }
    }
  }
}